Introduce ZoneWindowDrawing files and move stuff to better place (#6142)
