+++ src/modules/fancyzones/lib/FancyZonesLib.vcxproj
@@ -123,6 +123,7 @@
     <ClInclude Include="Zone.h" />
     <ClInclude Include="ZoneSet.h" />
     <ClInclude Include="ZoneWindow.h" />
+    <ClInclude Include="ZoneWindowDrawing.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="FancyZones.cpp" />
@@ -143,6 +144,7 @@
     <ClCompile Include="Zone.cpp" />
     <ClCompile Include="ZoneSet.cpp" />
     <ClCompile Include="ZoneWindow.cpp" />
+    <ClCompile Include="ZoneWindowDrawing.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="fancyzones.base.rc" />
