+++ src/modules/fancyzones/lib/FancyZonesLib.vcxproj.filters
@@ -74,6 +74,9 @@
     <ClInclude Include="KeyState.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="ZoneWindowDrawing.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp">
@@ -122,6 +125,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="FancyZonesDataTypes.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="ZoneWindowDrawing.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
